Prism Rewards
Updated the e2e test scripts
Updated both the dev and prod infra (we should now see a noticeable drop in cloud costs)
Rationalised the limit-order-mining (LOM) for order size, distance from market price and duration on the orderbook
Added new API endpoints to support this

Deprecated the old CreateMarket rpc service
Updated prism-docs

The Prism rewards accumulate hourly until we make them redeemable (e.g. end of season).

+++ api/server/lib/types.go
@@ -39,11 +39,11 @@ const (
 // price/distance/duration tiers:
 type PriceDistanceTier struct {
 	Threshold float64 // percent distance from market price (e.g., 0.20 for 20%)
-	Weight    int
+	Weight    float64
 }
 type OrderDurationTier struct {
 	Duration time.Duration
-	Weight   int
+	Weight   float64
 }
 
 // Zap logger setup
